Refactored digital device twin base classes into init.pys, moves is_equal and removing unneccessary imports
