+++ simpa/core/simulation_modules/__init__.py
@@ -9,7 +9,8 @@ from simpa.utils import Settings
 
 class SimulationModule(PipelineModule):
     """
-    Defines a simulation module that is a step in the simulation pipeline.
+    Defines a simulation module that is a step in the simulation pipeline. 
+    Each simulation module can only be one of Volume Creation, Light Propagation Modeling, Acoustic Wave Propagation Modeling, Image Reconstruction.
     """
 
     def __init__(self, global_settings: Settings):
@@ -28,17 +29,3 @@ class SimulationModule(PipelineModule):
         :return: Loads component settings corresponding to this simulation component
         """
         pass
-        
-    def before_running(self):
-        """
-        Called before running the run method from PipelineModule
-        """
-        self.logger.info(f"Starts running {self.__class__.__name__}")
-
-    def after_running(self):
-        """
-        Called after running the run method from PipelineModule
-        """
-        self.logger.info(f"Ended running {self.__class__.__name__}")
-
-
